+++ drift_dev/lib/src/writer/tables/data_class_writer.dart
@@ -138,9 +138,21 @@ class DataClassWriter {
     for (final column in columns) {
       final getter = column.dartGetterName;
       final jsonKey = column.getJsonKey(scope.options);
-      final type = column.dartTypeCode(scope.generationOptions);
+      String deserialized;
 
-      _buffer.write("$getter: serializer.fromJson<$type>(json['$jsonKey']),");
+      final typeConverter = column.typeConverter;
+      if (typeConverter != null && typeConverter.alsoAppliesToJsonConversion) {
+        final type = column.innerColumnType(scope.generationOptions);
+        final fromConverter = "serializer.fromJson<$type>(json['$jsonKey'])";
+        deserialized =
+            '${typeConverter.tableAndField}.fromJson($fromConverter)';
+      } else {
+        final type = column.dartTypeCode(scope.generationOptions);
+
+        deserialized = "serializer.fromJson<$type>(json['$jsonKey'])";
+      }
+
+      _buffer.write('$getter: $deserialized,');
     }
 
     _buffer.write(');}\n');
@@ -165,8 +177,14 @@ class DataClassWriter {
       final name = column.getJsonKey(scope.options);
       final getter = column.dartGetterName;
       final needsThis = getter == 'serializer';
-      final value = needsThis ? 'this.$getter' : getter;
-      final dartType = column.dartTypeCode(scope.generationOptions);
+      var value = needsThis ? 'this.$getter' : getter;
+      var dartType = column.dartTypeCode(scope.generationOptions);
+
+      final typeConverter = column.typeConverter;
+      if (typeConverter != null && typeConverter.alsoAppliesToJsonConversion) {
+        value = '${typeConverter.tableAndField}.toJson($value)';
+        dartType = '${column.innerColumnType(scope.generationOptions)}';
+      }
 
       _buffer.write("'$name': serializer.toJson<$dartType>($value),");
     }