+++ drift/CHANGELOG.md
@@ -4,6 +4,8 @@
   to define unique constraints in Dart tables.
 - Add the `check()` method to the Dart column builder to generate `CHECK` column
   constraints.
+- Also apply type converters for json serialization and deserialization if they
+  mix in `JsonTypeConverter`.
 - Add the very experimental `package:drift/wasm.dart` library. It uses WebAssembly
   to access sqlite3 without any external JavaScript libraries, but requires you to
   add a [WebAssembly module](https://github.com/simolus3/sqlite3.dart/tree/main/sqlite3#wasm-web-support)